backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy
pydantic-sqlalchemy
\nrequests
